requirements.txt
django
djangorestframework
python-dotenv
psycopg2-binary
drf-spectacular